Convert bandwith and min_size to int type

+++ cellulus/infer.py
@@ -19,10 +19,10 @@ def infer(experiment_config):
     inference_config = experiment_config.inference_config
 
     if inference_config.bandwidth is None:
-        inference_config.bandwidth = 0.5 * experiment_config.object_size
+        inference_config.bandwidth = int(0.5 * experiment_config.object_size)
 
     if inference_config.min_size is None:
-        inference_config.min_size = (
+        inference_config.min_size = int(
             0.1 * np.pi * (experiment_config.object_size**2) / 4
         )
